+++ engine/src/main/java/org/terasology/world/time/WorldTimeEvent.java
@@ -32,7 +32,7 @@ public class WorldTimeEvent extends TimeEventBase {
     public boolean matchesDaily(float fraction) {
         Preconditions.checkArgument(fraction >= 0 && fraction <= 1, "fraction must be in [0..1]");
 
-        long fracInMs = DoubleMath.roundToLong(fraction * WorldTime.DAY_LENGTH, RoundingMode.HALF_UP);
+        long fracInMs = DoubleMath.roundToLong((double) fraction * WorldTime.DAY_LENGTH, RoundingMode.HALF_UP);
         long diff = getDayTimeInMs() - fracInMs;
 
         return 2 * diff < WorldTime.TICK_EVENT_RATE && 2 * diff >= -WorldTime.TICK_EVENT_RATE;
